init some player of sd and refactor some logics ,also get player form database.

+++ BasketballScoutAndroidProject/BasketballScount/app/src/main/java/porcomsci/basketballscout/com/basketballscount/ManageSchoolInfoActivity.java
@@ -133,14 +133,18 @@ public class ManageSchoolInfoActivity extends ActionBarActivity {
                     AlertDialog editSchoolNameDialog = setUpEditSchoolNameDialog();
                     editSchoolNameDialog.show();
                 }
-                else // delete
+                else if(which==1) // delete
                 {
                     try {
                         deleteSchoolName();
                     } catch (SQLException e) {
                         e.printStackTrace();
                     }
-                }
+                }else{
+
+
+
+                }//manage player
 
             }
         });
